+++ test/e2e/cypress/integration/profiles/email/login/ui.spec.ts
@@ -31,7 +31,7 @@ context("UI tests using the email profile", () => {
       it("should use the json schema titles", () => {
         cy.get(`${appPrefix(app)}input[name="identifier"]`)
           .parent()
-          .should("contain.text", "ID")
+          .should("contain.text", "Your E-Mail")
 
         cy.get('input[name="password"]')
           .parentsUntil("label")